Removing -mx32 from the multilib list fixees most multilib problems.
It turns out that x86 & x86_64 need to be built with multilib (at
least to make fugu happy).  But they can't build with -mx32.
Removing -mx32 allows x86 & x86_64 to build happily with multilib.

Note: This CL depends on the changes to build-gcc.sh in
https://android-review.googlesource.com/329549

Bug: None
Test: Tested with toolchain/gcc/build.py & using the resulting prebuilts
to build angler, bullhead and fugu platforms.

Change-Id: If70a7b753913faa5ef271534a743a244b008c706